refactor: #52-folder-list 폴더 DTO/커맨드 네이밍 정리 및 파일명 리네임
