+++ src/main/java/com/keepgoing/keepgoing/folder/controller/dto/FolderCreateResponse.java
@@ -1,6 +1,6 @@
 package com.keepgoing.keepgoing.folder.controller.dto;
 
-import com.keepgoing.keepgoing.folder.service.dto.FolderResult;
+import com.keepgoing.keepgoing.folder.service.dto.FolderSummaryResult;
 
 public record CreateFolderResponse(
 		Long folderId,
@@ -8,7 +8,7 @@ public record CreateFolderResponse(
 		String name
 ) {
 
-	public static CreateFolderResponse from(FolderResult result) {
+	public static CreateFolderResponse from(FolderSummaryResult result) {
 		return new CreateFolderResponse(
 				result.folderId(),
 				result.parentFolderId(),
