+++ src/main/java/com/keepgoing/keepgoing/folder/service/dto/FolderSummaryResult.java
